requirements.txt
Flask
groq
python-dotenv
mysql-connector-python